split up ops tests into several files for readability
